requirements.txt
basestation
paho-mqtt
loguru
json5
python-dotenv
